refactor: move app components to UI
