+++ packages/cryptoplease/integration_test/app_test.dart
@@ -1,8 +1,8 @@
-import 'package:cryptoplease/app/components/recovery_phrase_text_view.dart';
 import 'package:cryptoplease/features/onboarding/presentation/sign_up/create_profile_screen.dart';
 import 'package:cryptoplease/features/onboarding/presentation/sign_up/get_started_screen.dart';
 import 'package:cryptoplease/main.dart' as app;
 import 'package:cryptoplease/ui/bottom_button.dart';
+import 'package:cryptoplease/ui/recovery_phrase_text_view.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
