+++ packages/cryptoplease/lib/app/screens/authenticated/profile/components/help_bottom_sheet.dart
@@ -1,8 +1,8 @@
-import 'package:cryptoplease/app/components/snackbars.dart';
 import 'package:cryptoplease/core/presentation/utils.dart';
 import 'package:cryptoplease/l10n/l10n.dart';
 import 'package:cryptoplease/ui/button.dart';
 import 'package:cryptoplease/ui/colors.dart';
+import 'package:cryptoplease/ui/snackbar.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
