+++ packages/cryptoplease/lib/app/screens/authenticated/profile/components/profile_section.dart
@@ -1,5 +1,4 @@
 import 'package:auto_route/auto_route.dart';
-import 'package:cryptoplease/app/components/dialogs.dart';
 import 'package:cryptoplease/app/routes.dart';
 import 'package:cryptoplease/app/screens/authenticated/profile/components/profile_button.dart';
 import 'package:cryptoplease/config.dart';
@@ -14,6 +13,7 @@ import 'package:cryptoplease/features/payment_request/bl/repository.dart';
 import 'package:cryptoplease/gen/assets.gen.dart';
 import 'package:cryptoplease/l10n/l10n.dart';
 import 'package:cryptoplease/ui/colors.dart';
+import 'package:cryptoplease/ui/dialogs.dart';
 import 'package:flutter/material.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:provider/provider.dart';
