+++ packages/cryptoplease/lib/app/screens/authenticated/profile/edit_profile_screen.dart
@@ -1,11 +1,11 @@
 import 'dart:io';
 
-import 'package:cryptoplease/app/components/decorated_window.dart';
 import 'package:cryptoplease/core/accounts/bl/accounts_bloc.dart';
 import 'package:cryptoplease/features/onboarding/presentation/sign_up/components/pick_profile_picture.dart';
 import 'package:cryptoplease/l10n/l10n.dart';
 import 'package:cryptoplease/ui/bottom_button.dart';
 import 'package:cryptoplease/ui/content_padding.dart';
+import 'package:cryptoplease/ui/decorated_window/decorated_window.dart';
 import 'package:cryptoplease/ui/loader.dart';
 import 'package:cryptoplease/ui/text_field.dart';
 import 'package:flutter/material.dart';
