+++ packages/cryptoplease/lib/app/screens/authenticated/wallet_flow/wallet_main_screen.dart
@@ -1,5 +1,3 @@
-import 'package:cryptoplease/app/components/number_formatter.dart';
-import 'package:cryptoplease/app/components/usdc_info.dart';
 import 'package:cryptoplease/core/amount.dart';
 import 'package:cryptoplease/features/outgoing_direct_payments/presentation/odp_header.dart';
 import 'package:cryptoplease/gen/assets.gen.dart';
@@ -8,6 +6,8 @@ import 'package:cryptoplease/l10n/l10n.dart';
 import 'package:cryptoplease/ui/amount_keypad/amount_keypad.dart';
 import 'package:cryptoplease/ui/button.dart';
 import 'package:cryptoplease/ui/navigation_bar/navigation_bar.dart';
+import 'package:cryptoplease/ui/number_formatter.dart';
+import 'package:cryptoplease/ui/usdc_info.dart';
 import 'package:decimal/decimal.dart';
 import 'package:flutter/material.dart';
