--- packages/cryptoplease/lib/app/text_styles.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-const twelveWordsTextStyle = TextStyle(
-  fontWeight: FontWeight.bold,
-  fontSize: 17,
-  color: Colors.white,
-  letterSpacing: 0.09,
-  wordSpacing: 2,
-  height: 1.411,
-);
