+++ packages/cryptoplease/lib/features/app_lock/presentation/app_lock_disable_screen.dart
@@ -1,8 +1,8 @@
-import 'package:cryptoplease/app/components/decorated_window.dart';
 import 'package:cryptoplease/features/app_lock/bl/app_lock_bloc.dart';
 import 'package:cryptoplease/features/app_lock/presentation/app_lock_setup_flow_screen.dart';
 import 'package:cryptoplease/features/app_lock/presentation/components/pin_input_display_widget.dart';
 import 'package:cryptoplease/l10n/l10n.dart';
+import 'package:cryptoplease/ui/decorated_window/decorated_window.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
