+++ packages/cryptoplease/lib/features/backup_phrase/presentation/backup_phrase_flow/backup_phrase_screen.dart
@@ -1,10 +1,10 @@
-import 'package:cryptoplease/app/components/decorated_window.dart';
-import 'package:cryptoplease/app/components/recovery_phrase_text_view.dart';
 import 'package:cryptoplease/core/accounts/bl/account.dart';
 import 'package:cryptoplease/features/backup_phrase/presentation/backup_phrase_flow/backup_phrase_flow_screen.dart';
 import 'package:cryptoplease/l10n/l10n.dart';
 import 'package:cryptoplease/ui/button.dart';
 import 'package:cryptoplease/ui/content_padding.dart';
+import 'package:cryptoplease/ui/decorated_window/decorated_window.dart';
+import 'package:cryptoplease/ui/recovery_phrase_text_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
