+++ packages/cryptoplease/lib/features/backup_phrase/presentation/puzzle_reminder/components/puzzle_screen.dart
@@ -1,5 +1,5 @@
-import 'package:cryptoplease/app/components/decorated_window_header.dart';
 import 'package:cryptoplease/ui/app_bar.dart';
+import 'package:cryptoplease/ui/decorated_window/decorated_window_header.dart';
 import 'package:cryptoplease/ui/theme.dart';
 import 'package:flutter/material.dart';
