+++ packages/cryptoplease/lib/features/incoming_split_key_payments/presentation/incoming_split_key_payment_screen.dart
@@ -1,10 +1,11 @@
-import 'package:cryptoplease/app/components/transfer_status/transfer_error.dart';
-import 'package:cryptoplease/app/components/transfer_status/transfer_progress.dart';
-import 'package:cryptoplease/app/components/transfer_status/transfer_success.dart';
+import 'package:auto_route/auto_route.dart';
 import 'package:cryptoplease/di.dart';
 import 'package:cryptoplease/features/incoming_split_key_payments/bl/incoming_split_key_payment.dart';
 import 'package:cryptoplease/features/incoming_split_key_payments/bl/iskp_bloc.dart';
 import 'package:cryptoplease/features/incoming_split_key_payments/bl/iskp_repository.dart';
+import 'package:cryptoplease/ui/transfer_status/transfer_error.dart';
+import 'package:cryptoplease/ui/transfer_status/transfer_progress.dart';
+import 'package:cryptoplease/ui/transfer_status/transfer_success.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -43,8 +44,11 @@ class _IncomingSplitKeyPaymentScreenState
               if (state.contains(payment.id)) return const TransferProgress();
 
               return payment.status.maybeMap(
-                success: (_) => const TransferSuccess(),
+                success: (_) => TransferSuccess(
+                  onOkPressed: () => context.router.pop(),
+                ),
                 orElse: () => TransferError(
+                  onBack: () => context.router.pop(),
                   onRetry: () => context
                       .read<ISKPBloc>()
                       .add(ISKPEvent.process(payment.id)),
