+++ packages/cryptoplease/lib/features/onboarding/presentation/login/restore_account_screen.dart
@@ -1,10 +1,10 @@
-import 'package:cryptoplease/app/components/onboarding_screen.dart';
-import 'package:cryptoplease/app/text_styles.dart';
 import 'package:cryptoplease/features/onboarding/bl/sign_up_bloc.dart';
 import 'package:cryptoplease/features/onboarding/presentation/login/mnemonic_input_formatter.dart';
 import 'package:cryptoplease/features/onboarding/presentation/sign_up/sign_up_flow_screen.dart';
 import 'package:cryptoplease/l10n/l10n.dart';
 import 'package:cryptoplease/ui/colors.dart';
+import 'package:cryptoplease/ui/onboarding_screen.dart';
+import 'package:cryptoplease/ui/theme.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
