+++ packages/cryptoplease/lib/features/onboarding/presentation/sign_up/create_profile_screen.dart
@@ -1,11 +1,11 @@
 import 'dart:io';
 
-import 'package:cryptoplease/app/components/dialogs.dart';
-import 'package:cryptoplease/app/components/onboarding_screen.dart';
 import 'package:cryptoplease/features/onboarding/bl/sign_up_bloc.dart';
 import 'package:cryptoplease/features/onboarding/presentation/sign_up/components/pick_profile_picture.dart';
 import 'package:cryptoplease/l10n/l10n.dart';
+import 'package:cryptoplease/ui/dialogs.dart';
 import 'package:cryptoplease/ui/loader.dart';
+import 'package:cryptoplease/ui/onboarding_screen.dart';
 import 'package:cryptoplease/ui/text_field.dart';
 import 'package:dfunc/dfunc.dart';
 import 'package:flutter/material.dart';
