+++ packages/cryptoplease/lib/features/onboarding/presentation/sign_up/create_recovery_phrase_screen.dart
@@ -1,8 +1,8 @@
-import 'package:cryptoplease/app/components/onboarding_screen.dart';
-import 'package:cryptoplease/app/components/recovery_phrase_text_view.dart';
 import 'package:cryptoplease/features/onboarding/bl/sign_up_bloc.dart';
 import 'package:cryptoplease/features/onboarding/presentation/sign_up/sign_up_flow_screen.dart';
 import 'package:cryptoplease/l10n/l10n.dart';
+import 'package:cryptoplease/ui/onboarding_screen.dart';
+import 'package:cryptoplease/ui/recovery_phrase_text_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
