+++ packages/cryptoplease/lib/features/onboarding/presentation/sign_up/no_email_and_password_screen.dart
@@ -1,8 +1,8 @@
-import 'package:cryptoplease/app/components/onboarding_screen.dart';
 import 'package:cryptoplease/features/onboarding/presentation/sign_up/sign_up_flow_screen.dart';
 import 'package:cryptoplease/l10n/l10n.dart';
 import 'package:cryptoplease/ui/info_icon.dart';
 import 'package:cryptoplease/ui/info_widget.dart';
+import 'package:cryptoplease/ui/onboarding_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:provider/provider.dart';
