+++ packages/cryptoplease/lib/features/outgoing_direct_payments/presentation/odp_confirmation_screen.dart
@@ -1,7 +1,4 @@
 import 'package:auto_route/auto_route.dart';
-import 'package:cryptoplease/app/components/dialogs.dart';
-import 'package:cryptoplease/app/components/number_formatter.dart';
-import 'package:cryptoplease/app/components/usdc_info.dart';
 import 'package:cryptoplease/core/tokens/token.dart';
 import 'package:cryptoplease/features/outgoing_direct_payments/presentation/odp_header.dart';
 import 'package:cryptoplease/l10n/device_locale.dart';
@@ -10,7 +7,10 @@ import 'package:cryptoplease/ui/amount_keypad/amount_keypad.dart';
 import 'package:cryptoplease/ui/button.dart';
 import 'package:cryptoplease/ui/chip.dart';
 import 'package:cryptoplease/ui/colors.dart';
+import 'package:cryptoplease/ui/dialogs.dart';
+import 'package:cryptoplease/ui/number_formatter.dart';
 import 'package:cryptoplease/ui/theme.dart';
+import 'package:cryptoplease/ui/usdc_info.dart';
 import 'package:decimal/decimal.dart';
 import 'package:dfunc/dfunc.dart';
 import 'package:flutter/material.dart';
