+++ packages/cryptoplease/lib/features/outgoing_direct_payments/presentation/odp_details_screen.dart
@@ -1,10 +1,11 @@
-import 'package:cryptoplease/app/components/transfer_status/transfer_error.dart';
-import 'package:cryptoplease/app/components/transfer_status/transfer_progress.dart';
-import 'package:cryptoplease/app/components/transfer_status/transfer_success.dart';
+import 'package:auto_route/auto_route.dart';
 import 'package:cryptoplease/di.dart';
 import 'package:cryptoplease/features/outgoing_direct_payments/bl/bloc.dart';
 import 'package:cryptoplease/features/outgoing_direct_payments/bl/outgoing_direct_payment.dart';
 import 'package:cryptoplease/features/outgoing_direct_payments/bl/repository.dart';
+import 'package:cryptoplease/ui/transfer_status/transfer_error.dart';
+import 'package:cryptoplease/ui/transfer_status/transfer_progress.dart';
+import 'package:cryptoplease/ui/transfer_status/transfer_success.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -41,8 +42,11 @@ class _ODPDetailsScreenState extends State<ODPDetailsScreen> {
               if (state.contains(payment.id)) return const TransferProgress();
 
               return payment.status.maybeMap(
-                success: (_) => const TransferSuccess(),
+                success: (_) => TransferSuccess(
+                  onOkPressed: () => context.router.pop(),
+                ),
                 orElse: () => TransferError(
+                  onBack: () => context.router.pop(),
                   onRetry: () =>
                       context.read<ODPBloc>().add(ODPEvent.process(payment.id)),
                 ),
