+++ packages/cryptoplease/lib/features/payment_request/presentation/link_details/tx_result_screen.dart
@@ -1,9 +1,9 @@
-import 'package:cryptoplease/app/components/decorated_window.dart';
 import 'package:cryptoplease/core/transactions/create_transaction_link.dart';
 import 'package:cryptoplease/gen/assets.gen.dart';
 import 'package:cryptoplease/l10n/l10n.dart';
 import 'package:cryptoplease/ui/colors.dart';
 import 'package:cryptoplease/ui/content_padding.dart';
+import 'package:cryptoplease/ui/decorated_window/decorated_window.dart';
 import 'package:cryptoplease/ui/theme.dart';
 import 'package:flutter/material.dart';
 import 'package:solana/solana.dart';
