+++ packages/cryptoplease/lib/ui/decorated_window/decorated_window.dart
@@ -1,6 +1,6 @@
-import 'package:cryptoplease/app/components/decorated_window_header.dart';
 import 'package:cryptoplease/ui/app_bar.dart';
 import 'package:cryptoplease/ui/background_gradient.dart';
+import 'package:cryptoplease/ui/decorated_window/decorated_window_header.dart';
 import 'package:cryptoplease/ui/theme.dart';
 import 'package:flutter/material.dart';
