+++ packages/cryptoplease/lib/ui/number_formatter.dart
