+++ packages/cryptoplease/lib/ui/snackbar.dart
@@ -1,3 +1,4 @@
+import 'package:cryptoplease/l10n/l10n.dart';
 import 'package:flutter/material.dart';
 
 void showCpSnackbar(
@@ -38,4 +39,10 @@ void showCpErrorSnackbar(BuildContext context, {required String message}) =>
       context,
       message: message,
       icon: const Icon(Icons.error, color: Colors.red),
+    );
+
+void showClipboardSnackbar(BuildContext context) => showCpSnackbar(
+      context,
+      message: context.l10n.copiedToClipboard,
+      icon: const Icon(Icons.check, color: Colors.green),
     );
