+++ packages/cryptoplease/lib/ui/theme.dart
@@ -151,3 +151,12 @@ class CpThemeData {
         dividerColor: dividerColor,
       );
 }
+
+const twelveWordsTextStyle = TextStyle(
+  fontWeight: FontWeight.bold,
+  fontSize: 17,
+  color: Colors.white,
+  letterSpacing: 0.09,
+  wordSpacing: 2,
+  height: 1.411,
+);
