+++ packages/cryptoplease/lib/ui/transfer_status/transfer_success.dart
@@ -1,37 +1,33 @@
-import 'package:auto_route/auto_route.dart';
 import 'package:cryptoplease/l10n/l10n.dart';
 import 'package:cryptoplease/ui/button.dart';
 import 'package:cryptoplease/ui/status_screen.dart';
 import 'package:cryptoplease/ui/status_widget.dart';
 import 'package:flutter/material.dart';
 
-class TransferError extends StatelessWidget {
-  const TransferError({
+class TransferSuccess extends StatelessWidget {
+  const TransferSuccess({
     super.key,
-    required this.onRetry,
+    required this.onOkPressed,
   });
 
-  final VoidCallback onRetry;
+  final VoidCallback onOkPressed;
 
   @override
   Widget build(BuildContext context) => StatusScreen(
         title: context.l10n.splitKeyTransferTitle,
-        statusTitle: Text(context.l10n.splitKeyErrorMessage1),
-        onBackButtonPressed: () => context.router.pop(),
-        statusContent: Text(
-          '${context.l10n.splitKeyErrorMessage2} ${context.l10n.splitKeyErrorRetry}',
-        ),
-        statusType: CpStatusType.error,
+        statusTitle: Text(context.l10n.splitKeySuccessMessage1),
+        statusContent: Text(context.l10n.splitKeySuccessMessage2),
+        statusType: CpStatusType.success,
         content: Padding(
           padding: const EdgeInsets.symmetric(horizontal: 24),
           child: Column(
             children: [
-              const SizedBox(height: 140),
+              const SizedBox(height: 160),
               CpButton(
                 size: CpButtonSize.big,
                 width: double.infinity,
-                text: context.l10n.retry,
-                onPressed: onRetry,
+                text: context.l10n.ok,
+                onPressed: onOkPressed,
               )
             ],
           ),
