+++ packages/cryptoplease/lib/ui/usdc_info.dart
